Unify plotting scripts
